Conflicts:
	modules/ballerina-core/src/main/java/org/ballerinalang/model/types/BDoubleType.java
	modules/ballerina-core/src/main/java/org/ballerinalang/model/types/BLongType.java
	modules/ballerina-core/src/main/resources/MessagesBundle.properties
	modules/ballerina-core/src/main/resources/MessagesBundle_en_US.properties

+++ modules/ballerina-core/src/main/java/org/ballerinalang/bre/BLangExecutor.java
@@ -413,7 +413,7 @@ public class BLangExecutor implements NodeExecutor {
         BMessage inMsg = (BMessage) expr.execute(this);
         List<WorkerRunner> workerRunnerList = new ArrayList<>();
         List<BMessage> resultMsgs = new ArrayList<>();
-        int timeout = ((BInteger) forkJoinStmt.getTimeout().getTimeoutExpression().execute(this)).intValue();
+        long timeout = ((BInteger) forkJoinStmt.getTimeout().getTimeoutExpression().execute(this)).intValue();
 
         Worker[] workers = forkJoinStmt.getWorkers();
         Map<String, WorkerRunner> triggeredWorkers = new HashMap<>();
@@ -518,7 +518,7 @@ public class BLangExecutor implements NodeExecutor {
 
     }
 
-    private BMessage invokeAnyWorker(List<WorkerRunner> workerRunnerList, int timeout) {
+    private BMessage invokeAnyWorker(List<WorkerRunner> workerRunnerList, long timeout) {
         ExecutorService anyExecutor = Executors.newWorkStealingPool();
         BMessage result;
         try {
@@ -532,7 +532,7 @@ public class BLangExecutor implements NodeExecutor {
         return result;
     }
 
-    private List<BMessage> invokeAllWorkers(List<WorkerRunner> workerRunnerList, int timeout) {
+    private List<BMessage> invokeAllWorkers(List<WorkerRunner> workerRunnerList, long timeout) {
         ExecutorService allExecutor = Executors.newWorkStealingPool();
         List<BMessage> result = new ArrayList<>();
         try {
